pyproject.toml
[tool.poetry]
name = "django-doma"
version = "0.1.0"
description = "Simple Document Management for Django"
authors = ["Florian Rämisch <olf@subsignal.org>"]
license = "GPL-3.0"

packages = [
    { include="doma", from="." },
]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
black = "^20.8b1"
pre-commit = "^2.12.0"
factory-boy = "^3.2.0"
poetry = "^1.1.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
